+++ src/instanceinator.ts
@@ -5,6 +5,16 @@ import { copyInstance, type InstanceinatorCopyInstanceConfig } from './instance-
 import { FpsLabelDrawClass, IdLabelDrawClass, type LabelDrawClass } from './label-draw'
 import { updateMusicInstanceId } from './fixes/music-fix'
 import { classes } from './custom-classes'
+import {
+    wrap,
+    scheduleNextTask,
+    schedulePostTask,
+    scheduleTask,
+    scheduleTasks,
+    runTask,
+    runTasks,
+    wait,
+} from './inst-util'
 
 declare global {
     var instanceinator: Instanceinator
@@ -74,4 +84,13 @@ export class Instanceinator {
     retile = retile
     copy = copyInstance
     classes!: typeof classes
+
+    scheduleTask = scheduleTask
+    schedulePostTask = schedulePostTask
+    scheduleNextTask = scheduleNextTask
+    scheduleTasks = scheduleTasks
+    wrap = wrap
+    runTask = runTask
+    runTasks = runTasks
+    wait = wait
 }